Always unmounting dmg file.

+++ scripts/sign_macos_packages.sh
@@ -241,34 +241,8 @@ sign_zip_content () {
 	return 0
 }
 
-# Brief create temporary file copy.
-tmp_file_copy () {
-	local ORIG_FILE="$1"
-	if [ "x${ORIG_FILE}" = "x" ]; then
-		echo "Expected non-empty file name." >&2
-		return 1
-	fi
-
-	if [ ! -f "${ORIG_FILE}" ]; then
-		echo "'${ORIG_FILE}' is not a file." >&2
-		return 1
-	fi
-
-	local FILE_NAME=$(file_name ${ORIG_FILE})
-	FILE_NAME=$(echo ${FILE_NAME} | sed -e 's/[.]/_/g')
-	local TMP_FILE=$(mktemp "/tmp/tmp_${FILE_NAME}_XXXX")
-	if [ "x${TMP_FILE}" = "x" ]; then
-		echo "Cannot create temporary file." >&2
-		return 1
-	fi
-
-	cp "${ORIG_FILE}" "${TMP_FILE}" || return 1
-
-	echo "${TMP_FILE}"
-	return 0
-}
-
-remove_symlinks () {
+# Return all files that are not symlinks.
+exclude_symlinks () {
 	local FILE_LIST="$@"
 	for FILE in ${FILE_LIST}; do
 		if [ -e "${FILE}" -a ! -L "${FILE}" ]; then
@@ -277,33 +251,24 @@ remove_symlinks () {
 	done
 }
 
-# Sign content of dmg package and the package itself.
-sign_dmg () {
-	local UNSIGNED_DMG="$1"
-	if [ "x${UNSIGNED_DMG}" = "x" ]; then
-		echo "Expected non-empty dmg file name." >&2
+# Sign dmg package content that has already been mounter into a directory.
+sign_dmg_content () {
+	local VOLUME_DIR="$1"
+	if [ "x${VOLUME_DIR}" = "x" ]; then
+		echo "Expected non-empty volume directory." >&2
 		return 1
 	fi
 
-	UNSIGNED_DMG=$(realpath "${UNSIGNED_DMG}")
-	local TMP_DMG=$(tmp_file_copy "${UNSIGNED_DMG}") # Temporary copy.
-	local SIGNED_DMG=$(signed_absolute_file_path "${UNSIGNED_DMG}")
-
-	echo "'${UNSIGNED_DMG}' -> '${SIGNED_DMG}'"
-
-	echo "Opening dmg package '${TMP_DMG}'"
-	local DEVICE=$(hdiutil attach -owners on "${TMP_DMG}" -shadow | egrep '^/dev/' | sed 1q | awk '{print $1}')
-	if [ "x${DEVICE}" = "x" ]; then
-		echo "Error opening dmg package '${TMP_DMG}'." >&2
+	if [ ! -d "${VOLUME_DIR}" ]; then
+		echo "'${VOLUME_DIR}' is not a directory." >&2
 		return 1
 	fi
-	echo "Dmg package opened as device '${DEVICE}'."
 
 	pushd "${VOLUME_DIR}" > /dev/null
 
 	local APP_ROOT="${APP_NAME}.app"
 	local DYLIB_FILES=$(insensitive_match_suffix "dylib" $(find "${APP_ROOT}"))
-	DYLIB_FILES=$(remove_symlinks ${DYLIB_FILES})
+	DYLIB_FILES=$(exclude_symlinks ${DYLIB_FILES})
 	local FRAMEWORKS=""
 	FRAMEWORKS="${FRAMEWORKS} ${APP_ROOT}/Contents/Frameworks/QtCore.framework/Versions/5/QtCore"
 	FRAMEWORKS="${FRAMEWORKS} ${APP_ROOT}/Contents/Frameworks/QtGui.framework/Versions/5/QtGui"
@@ -319,13 +284,46 @@ sign_dmg () {
 	# Sign the whole application.
 	${SIGN_CMD} "${SIGN_CERT_ID}" "${APP_ROOT}" || return 1
 
-	codesign -dv --verbose=4 "${APP_ROOT}" || return 1
+	${SIGN_VERIFY_CMD} "${APP_ROOT}" || return 1
 
 	popd > /dev/null
 
+	return 0
+}
+
+# Sign content of dmg package and the package itself.
+sign_dmg () {
+	local UNSIGNED_DMG="$1"
+	if [ "x${UNSIGNED_DMG}" = "x" ]; then
+		echo "Expected non-empty dmg file name." >&2
+		return 1
+	fi
+
+	UNSIGNED_DMG=$(realpath "${UNSIGNED_DMG}")
+	local SIGNED_DMG=$(signed_absolute_file_path "${UNSIGNED_DMG}")
+
+	local TMP_DIR=$(mktemp -d /tmp/sign_XXXX)
+	local TMP_DMG=$(file_name "${UNSIGNED_DMG}")
+	TMP_DMG="${TMP_DIR}/${TMP_DMG}"
+	cp "${UNSIGNED_DMG}" "${TMP_DMG}" || exit 1 # Temporary copy.
+
+	echo "'${UNSIGNED_DMG}' -> '${SIGNED_DMG}'"
+
+	echo "Opening dmg package '${TMP_DMG}'"
+	local DEVICE=$(hdiutil attach -owners on "${TMP_DMG}" -shadow | egrep '^/dev/' | sed 1q | awk '{print $1}')
+	if [ "x${DEVICE}" = "x" ]; then
+		echo "Error opening dmg package '${TMP_DMG}'." >&2
+		return 1
+	fi
+	echo "Dmg package opened as device '${DEVICE}'."
+
+	sign_dmg_content "${VOLUME_DIR}"
+
+	# The volume must be detached even on errors.
 	hdiutil detach "${DEVICE}"
 	hdiutil convert -format UDZO -o "${SIGNED_DMG}" "${TMP_DMG}" -shadow || return 1
 	rm "${TMP_DMG}" "${TMP_DMG}.shadow"
+	rm -r "${TMP_DIR}"
 
 	# Sign whole package.
 	${SIGN_CMD} "${SIGN_CERT_ID}" "${SIGNED_DMG}" || return 1
@@ -343,14 +341,20 @@ sign_win_installer () {
 	fi
 
 	UNSIGNED_INST=$(realpath "${UNSIGNED_INST}")
-	local TMP_INST=$(tmp_file_copy "${UNSIGNED_INST}") # Temporary copy.
 	local SIGNED_INST=$(signed_absolute_file_path "${UNSIGNED_INST}")
+
+	local TMP_DIR=$(mktemp -d /tmp/sign_XXXX)
+	local TMP_INST=$(file_name "${UNSIGNED_INST}")
+	TMP_INST="${TMP_DIR}/${TMP_INST}"
+	cp "${UNSIGNED_INST}" "${TMP_INST}" || exit 1 # Temporary copy.
 
 	echo "'${UNSIGNED_INST}' -> '${SIGNED_INST}'"
 
 	${SIGN_CMD} "${SIGN_CERT_ID}" "${TMP_INST}" || return 1
 	mv "${TMP_INST}" "${SIGNED_INST}"
 	${SIGN_VERIFY_CMD} "${SIGNED_INST}" || return 1
+
+	rm -r "${TMP_DIR}"
 
 	return 0
 }
